requirements.txt
flask>=2.3
numpy>=1.24
Pillow>=10.0
opencv-python-headless>=4.8
matplotlib>=3.7
gunicorn>=21.0
